Update 1129-unicode-general-category-number.html
Hello! Maybe I was wrong, but isn't this https://hydrocul.github.io/wiki/blog/2015/1129-unicode-general-category-number.html your article? 
I have a global task to replace unicode-table.com url to symbl.cc on all Github pages. This article contains a hundreds of incorrect links, and i need to fix it (( 
If I made a mistake and this is not your article, I apologize, my changes in your file are incorrect, I used file editing to get in touch with you))